+++ src/test/php/util/cmd/unittest/AbstractRunnerTest.class.php
@@ -70,9 +70,9 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
    */
   protected function newCommand() {
     return newinstance(Command::class, [], '{
-      public static $wasRun= FALSE;
-      public function __construct() { self::$wasRun= FALSE; }
-      public function run() { self::$wasRun= TRUE; }
+      public static $wasRun= false;
+      public function __construct() { self::$wasRun= false; }
+      public function run() { self::$wasRun= true; }
       public function wasRun() { return self::$wasRun; }
     }');
   }
@@ -81,7 +81,7 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   public function nonExistant() {
     $return= $this->runWith(['@@NONEXISTANT@@']);
     $this->assertEquals(1, $return);
-    $this->assertOnStream($this->err, '*** No command named "@@NONEXISTANT@@"');
+    $this->assertOnStream($this->err, '*** Class "@@NONEXISTANT@@" could not be found');
     $this->assertEquals('', $this->out->getBytes());
   }
 
@@ -154,7 +154,7 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function positionalArgument() {
     $command= newinstance(Command::class, [], '{
-      private $arg= NULL;
+      private $arg= null;
 
       #[@arg(position= 0)]
       public function setArg($arg) { $this->arg= $arg; }
@@ -170,7 +170,7 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function missingPositionalArgumentt() {
     $command= newinstance(Command::class, [], '{
-      private $arg= NULL;
+      private $arg= null;
 
       #[@arg(position= 0)]
       public function setArg($arg) { $this->arg= $arg; }
@@ -186,7 +186,7 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function shortNamedArgument() {
     $command= newinstance(Command::class, [], '{
-      private $arg= NULL;
+      private $arg= null;
 
       #[@arg]
       public function setArg($arg) { $this->arg= $arg; }
@@ -202,7 +202,7 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function longNamedArgument() {
     $command= newinstance(Command::class, [], '{
-      private $arg= NULL;
+      private $arg= null;
 
       #[@arg]
       public function setArg($arg) { $this->arg= $arg; }
@@ -218,7 +218,7 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function shortRenamedArgument() {
     $command= newinstance(Command::class, [], '{
-      private $arg= NULL;
+      private $arg= null;
 
       #[@arg(name= "pass")]
       public function setArg($arg) { $this->arg= $arg; }
@@ -234,7 +234,7 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function longRenamedArgument() {
     $command= newinstance(Command::class, [], '{
-      private $arg= NULL;
+      private $arg= null;
 
       #[@arg(name= "pass")]
       public function setArg($arg) { $this->arg= $arg; }
@@ -250,7 +250,7 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function missingNamedArgument() {
     $command= newinstance(Command::class, [], '{
-      private $arg= NULL;
+      private $arg= null;
 
       #[@arg]
       public function setArg($arg) { $this->arg= $arg; }
@@ -266,10 +266,10 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function existanceArgumentNotPassed() {
     $command= newinstance(Command::class, [], '{
-      private $verbose= FALSE;
+      private $verbose= false;
 
       #[@arg]
-      public function setVerbose() { $this->verbose= TRUE; }
+      public function setVerbose() { $this->verbose= true; }
       public function run() { $this->out->write($this->verbose ? "true" : "false"); }
     }');
 
@@ -282,8 +282,8 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function optionalArgument() {
     $command= newinstance(Command::class, [], '{
-      private $verbose= FALSE;
-      private $name= NULL;
+      private $verbose= false;
+      private $name= null;
 
       #[@arg]
       public function setName($name= "unknown") { $this->name= $name; }
@@ -299,8 +299,8 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function optionalArgumentNotPassed() {
     $command= newinstance(Command::class, [], '{
-      private $verbose= FALSE;
-      private $name= NULL;
+      private $verbose= false;
+      private $name= null;
 
       #[@arg]
       public function setName($name= "unknown") { $this->name= $name; }
@@ -316,10 +316,10 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function shortExistanceArgumentPassed() {
     $command= newinstance(Command::class, [], '{
-      private $verbose= FALSE;
+      private $verbose= false;
 
       #[@arg]
-      public function setVerbose() { $this->verbose= TRUE; }
+      public function setVerbose() { $this->verbose= true; }
       public function run() { $this->out->write($this->verbose ? "true" : "false"); }
     }');
 
@@ -332,10 +332,10 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function longExistanceArgumentPassed() {
     $command= newinstance(Command::class, [], '{
-      private $verbose= FALSE;
+      private $verbose= false;
 
       #[@arg]
-      public function setVerbose() { $this->verbose= TRUE; }
+      public function setVerbose() { $this->verbose= true; }
       public function run() { $this->out->write($this->verbose ? "true" : "false"); }
     }');
 
@@ -384,7 +384,7 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function allArgs() {
     $this->assertAllArgs('a, b, c, d, e, f, g', newinstance(Command::class, [], '{
-      private $verbose= FALSE;
+      private $verbose= false;
       private $args= [];
 
       #[@args(select= "[0..]")]
@@ -396,7 +396,7 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function allArgsCompactNotation() {
     $this->assertAllArgs('a, b, c, d, e, f, g', newinstance(Command::class, [], '{
-      private $verbose= FALSE;
+      private $verbose= false;
       private $args= [];
 
       #[@args(select= "*")]
@@ -408,7 +408,7 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function boundedArgs() {
     $this->assertAllArgs('a, b, c', newinstance(Command::class, [], '{
-      private $verbose= FALSE;
+      private $verbose= false;
       private $args= [];
 
       #[@args(select= "[0..2]")]
@@ -420,7 +420,7 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function boundedArgsFromOffset() {
     $this->assertAllArgs('c, d, e', newinstance(Command::class, [], '{
-      private $verbose= FALSE;
+      private $verbose= false;
       private $args= [];
 
       #[@args(select= "[2..4]")]
@@ -432,7 +432,7 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function positionalAndBoundedArgsFromOffset() {
     $this->assertAllArgs('a, c, d, e', newinstance(Command::class, [], '{
-      private $verbose= FALSE;
+      private $verbose= false;
       private $args= [];
 
       #[@args(select= "0, [2..4]")]
@@ -444,7 +444,7 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function boundedAndPositionalArgsWithOverlap() {
     $this->assertAllArgs('a, b, c, b', newinstance(Command::class, [], '{
-      private $verbose= FALSE;
+      private $verbose= false;
       private $args= [];
 
       #[@args(select= "[0..2], 1")]
@@ -456,7 +456,7 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function positionalArgs() {
     $this->assertAllArgs('a, c, e, f', newinstance(Command::class, [], '{
-      private $verbose= FALSE;
+      private $verbose= false;
       private $args= [];
 
       #[@args(select= "0, 2, 4, 5")]
@@ -468,11 +468,11 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
   #[@test]
   public function configOption() {
     $command= newinstance(Command::class, [], '{
-      private $choke= FALSE;
+      private $choke= false;
 
       #[@arg]
       public function setChoke() { 
-        $this->choke= TRUE; 
+        $this->choke= true; 
       }
       
       public function run() { 
@@ -484,6 +484,7 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
     $this->assertEquals('', $this->err->getBytes());
     $this->assertEquals('true', $this->out->getBytes());
   }
+
 
   #[@test]
   public function class_with_create() {
@@ -538,5 +539,16 @@ abstract class AbstractRunnerTest extends \unittest\TestCase {
     }');
     $this->runWith([nameof($command)]);
     $this->assertOnStream($this->out, 'Created with util.cmd.Config');
+  }
+
+  #[@test]
+  public function can_be_invoked_via_main() {
+    $command= newinstance(Command::class, [], [
+      'arg' => 0,
+      '#[@arg] setArg' => function($arg) { $this->arg= $arg; },
+      'run' => function() { return $this->arg; }
+    ]);
+
+    $this->assertEquals(12, $command::main(['-a', 12]));
   }
 }